move disk funcs
