+++ MANIFEST.in
@@ -63,6 +63,8 @@ include psutil/arch/freebsd/sys_socks.c
 include psutil/arch/freebsd/sys_socks.h
 include psutil/arch/netbsd/cpu.c
 include psutil/arch/netbsd/cpu.h
+include psutil/arch/netbsd/disk.c
+include psutil/arch/netbsd/disk.h
 include psutil/arch/netbsd/mem.c
 include psutil/arch/netbsd/mem.h
 include psutil/arch/netbsd/socks.c
