+++ psutil/_psutil_bsd.c
@@ -94,6 +94,7 @@
     #include <sys/sched.h>  // for CPUSTATES & CP_*
 #elif PSUTIL_NETBSD
     #include "arch/netbsd/cpu.h"
+    #include "arch/netbsd/disk.h"
     #include "arch/netbsd/mem.h"
     #include "arch/netbsd/specific.h"
     #include "arch/netbsd/socks.h"
