+++ setup.py
@@ -288,6 +288,7 @@ elif NETBSD:
         sources=sources + [
             'psutil/_psutil_bsd.c',
             'psutil/arch/netbsd/cpu.c',
+            'psutil/arch/netbsd/disk.c',
             'psutil/arch/netbsd/mem.c',
             'psutil/arch/netbsd/specific.c',
             'psutil/arch/netbsd/socks.c',
